More changes.

+++ bash_aliases.sh
@@ -9,6 +9,10 @@ alias -- -="cd -"
 # Enable sudo for aliases
 alias sudo='sudo '
 
+# Find PID of a process
+# Usage: ps? [name]
+alias ps?="ps aux | grep -i"
+
 # Detect ls for proper coloring
 if ls --color > /dev/null 2>&1; then # GNU ls
   colorflag="--color"
@@ -18,7 +22,6 @@ fi
 alias ls="ls -lAHFp ${colorflag}"
 export CLICOLOR=1
 export LS_COLORS='no=00:fi=00:di=01;34:ln=01;36:pi=40;33:so=01;35:do=01;35:bd=40;33;01:cd=40;33;01:or=40;31;01:ex=01;32:*.tar=01;31:*.tgz=01;31:*.arj=01;31:*.taz=01;31:*.lzh=01;31:*.zip=01;31:*.z=01;31:*.Z=01;31:*.gz=01;31:*.bz2=01;31:*.deb=01;31:*.rpm=01;31:*.jar=01;31:*.jpg=01;35:*.jpeg=01;35:*.gif=01;35:*.bmp=01;35:*.pbm=01;35:*.pgm=01;35:*.ppm=01;35:*.tga=01;35:*.xbm=01;35:*.xpm=01;35:*.tif=01;35:*.tiff=01;35:*.png=01;35:*.mov=01;35:*.mpg=01;35:*.mpeg=01;35:*.avi=01;35:*.fli=01;35:*.gl=01;35:*.dl=01;35:*.xcf=01;35:*.xwd=01;35:*.ogg=01;35:*.mp3=01;35:*.wav=01;35:'
-
 
 # IP addresses
 alias ip="dig +short myip.opendns.com @resolver1.opendns.com"
@@ -35,9 +38,11 @@ alias which='type -all' # better which
 alias gitpull='git pull origin master'
 alias gitpush='git push origin master'
 
-# Show/hide hidden files in Finder
-alias show="defaults write com.apple.finder AppleShowAllFiles -bool true && killall Finder"
-alias hide="defaults write com.apple.finder AppleShowAllFiles -bool false && killall Finder"
+if is_osx; then
+  # Show/hide hidden files in Finder
+  alias show="defaults write com.apple.finder AppleShowAllFiles -bool true && killall Finder"
+  alias hide="defaults write com.apple.finder AppleShowAllFiles -bool false && killall Finder"
+fi
 
 # Reload the shell (i.e. invoke as a login shell)
 alias reload="exec $SHELL -l"
